basicbot/requirements.txt
# BasicBot Trading System Requirements

# Core dependencies
pandas>=1.5.0
numpy>=1.22.0
python-dotenv>=0.19.0
requests>=2.27.0

# Trading API
alpaca-trade-api>=2.3.0
yfinance>=0.2.3

# Technical analysis
ta>=0.10.0               # Technical analysis indicators
pandas-ta>=0.3.14b0      # Additional technical indicators

# Data visualization
matplotlib>=3.5.0
seaborn>=0.11.0

# Logging and monitoring
loguru>=0.6.0

# Database
sqlalchemy>=1.4.0        # For SQLite database logging

# Utilities
tqdm>=4.62.0             # Progress bars
pytz>=2021.3             # Timezone utilities

# Trading and analysis libraries
scikit-learn>=1.0.2
tensorflow>=2.9.1
joblib>=1.1.0
matplotlib>=3.5.2

# Data handling
pytz>=2022.1

# Database connectors
psycopg2-binary>=2.9.3
mysql-connector-python>=8.0.29

# Testing and development
pytest>=7.1.2
pytest-cov>=3.0.0
black>=22.6.0
flake8>=4.0.1

# Utilities
PyYAML>=6.0 